+++ .gitignore
@@ -128,7 +128,9 @@ celerybeat.pid
 # Environments
 .env
 .venv
+env/
 venv/
+ENV/
 env.bak/
 venv.bak/
