+++ nanoswea/run_github_issue.py
@@ -6,7 +6,7 @@ import requests
 import yaml
 
 from nanoswea.agent import Agent, AgentConfig
-from nanoswea.env import DockerEnvironment
+from nanoswea.environment import DockerEnvironment
 from nanoswea.model import LitellmModel, ModelConfig
